+++ codon/parser/visitors/typecheck/assign.cpp
@@ -196,7 +196,7 @@ Stmt *TypecheckVisitor::transformAssignment(AssignStmt *stmt, bool mustExist) {
   bool isGlobal = (ctx->cache->isJit && val->isGlobal() && !val->isGeneric()) ||
                   (canonical == VAR_ARGV);
   if (isGlobal && val->isVar())
-    registerGlobal(canonical);
+    registerGlobal(canonical, assign->getRhs());
 
   return assign;
 }