+++ codon/parser/visitors/typecheck/typecheck.h
@@ -329,13 +329,14 @@ public:
     return t;
   }
   types::TypePtr instantiateType(types::Type *t);
-  void registerGlobal(const std::string &s);
+  void registerGlobal(const std::string &s, bool = false);
   types::ClassType *getStdLibType(const std::string &type);
   types::Type *extractClassGeneric(types::Type *t, int idx = 0);
   types::Type *extractFuncGeneric(types::Type *t, int idx = 0);
   types::Type *extractFuncArgType(types::Type *t, int idx = 0);
   std::string getClassMethod(types::Type *typ, const std::string &member);
   std::string getTemporaryVar(const std::string &s);
+  bool isImportFn(const std::string &s);
 
   int64_t getIntLiteral(types::Type *t, size_t pos = 0);
   bool getBoolLiteral(types::Type *t, size_t pos = 0);