+++ docs/intro/releases.md
@@ -18,6 +18,7 @@ Upgraded to LLVM 17 (from 15).
 - Updates to several existing functions, such as adding `key` and
   `default` arguments to `min()` and `max()`.
 - Slice arguments can now be of any type, not just `int`.
+- Added `input()` function.
 
 ## Other improvements
 
@@ -25,6 +26,13 @@ Upgraded to LLVM 17 (from 15).
 - Updated import logic to match CPython's more closely.
 - Several improvements to dynamic polymorphism to match CPython more
   closely.
+
+## New compiler options
+
+- `-disable-exceptions` will disable exceptions, potentially eliding
+  various runtime checks (e.g. bounds checks for lists). This flag
+  should only be used if you know that no exceptions will be raised
+  in the given program.
 
 # v0.16